requirements.txt
cryptography>=36.0.1
PyJWT>=2.3.0
tinydb>=4.5.2
flask-restful>=0.3.9
